# src/mw_postgrest/docs/backend-openapi.yaml
openapi: 3.0.0
info:
  title: CertM3 API
  description: |
    API for managing certificates, users, groups, and certificate requests.
    This API provides endpoints for certificate lifecycle management, user management,
    group management, and certificate request processing.

    Note: The /api prefix is added by nginx configuration. All endpoints are
    accessible without the /api prefix in the controllers.

    Architecture:
    - The API is designed to be used internally by the CertM3 application server
    - The app server acts as a proxy, handling JWT validation and session management
    - Browser clients communicate with the app server, not directly with the API
    - The API can be kept private/internal for security

    Special Endpoints:
    - /request/check-username/{username}: This endpoint uses the singular form
      to indicate it's a utility operation separate from the main request lifecycle.
      It's called frequently during username input and may have different
      optimization/rate-limiting requirements.

    Test Coverage:
    - /users/username/{username}: Covered by unit tests
    - Other endpoints: Covered by integration tests
  version: 1.0.0
  contact:
    name: ogt11.com/llc
servers:
  - url: https://urp.ogt11.com/api
    description: Production server
components:
  schemas:
    User:
      type: object
      required:
        - username
        - email
        - displayName
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user (auto-generated)
        username:
          type: string
          description: Unique username
        email:
          type: string
          format: email
          description: User's email address (unique)
        displayName:
          type: string
          description: User's display name (defaults to 'Unknown')
        status:
          type: string
          enum: [active, inactive]
          description: User's current status
        createdAt:
          type: string
          format: date-time
          description: When the user was created
        createdBy:
          type: string
          description: ID of the user who created this user
        updatedAt:
          type: string
          format: date-time
          description: When the user was last updated
        updatedBy:
          type: string
          description: ID of the user who last updated this user
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - statusCode
            - name
            - message
          properties:
            statusCode:
              type: integer
              description: HTTP status code
            name:
              type: string
              description: Error name
            message:
              type: string
              description: Error message
...
paths:
...
  /request/check-username/{username}:
    get:
      x-controller-name: RequestController
      x-operation-name: checkUsername
      tags:
        - RequestController
      summary: Check username availability
      description: Checks if a username is available for registration
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
          description: Username to check
      responses:
        '200':
          description: Username is available
        '404':
          description: Username is taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: RequestController.checkUsername
...
  /requests/{id}/validate:
    post:
      x-controller-name: RequestController
      x-operation-name: validate
      tags:
        - RequestController
      summary: Validate a request
      description: Validates a certificate request. On success, creates the user and returns their userId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Request ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - challenge
              properties:
                challenge:
                  type: string
                  description: Validation challenge token
      responses:
        '200':
          description: Request validated successfully and user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                    format: uuid
                    description: ID of the created user
        '400':
          description: Invalid challenge
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: RequestController.validate
...
  /users/{userId}/groups:
    get:
      x-controller-name: UserController
      x-operation-name: getUserGroups
      tags:
        - UserController
      summary: Get user's groups
      description: Retrieves all groups a user belongs to
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        '200':
          description: Array of group names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      operationId: UserController.getUserGroups
  /users/username/{username}:
    get:
      summary: Get user by username
      operationId: getUserByUsername
      tags:
        - users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            maxLength: 255
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'